Fixed Bugs, improved performance and reliability
Fixed minor bug in pdf_utils.py method 'clean_index_text'. in the regex pattern that caused it to remain in a never-ending cycle.

Fixed bug in pdf_meta_extract.py . When is not possible read the content of a page we try with the next one. At least a max number of times.

Fixed minor bug in pdf_meta_extract.py . When is not possible read the content of pdf at least we return the information we were able to get.

In logging_config.py we disable debug logs from validators library